requirements.txt
numpy==1.15.1
scipy==0.19
Pillow
cython
matplotlib
scikit-image
tensorflow==1.3.0
keras==2.0.8
opencv-python
h5py==2.7.0
imgaug==0.2.6
IPython[all]
pandas==0.20.3
seaborn==0.8
pywavelets==0.5.2
Django==2.1.4